fix(auth): guard protected pages via SSR
