+++ web/src/app/groups/[slug]/reservations/new/page.tsx
@@ -1,12 +1,15 @@
+export const dynamic = 'force-dynamic';
+export const revalidate = 0;
+export const fetchCache = 'default-no-store';
+
 import { serverFetch } from '@/lib/server-fetch';
 import { notFound, redirect } from 'next/navigation';
+import { getUserFromCookies } from '@/lib/auth/server';
 import NewReservationClient from './Client';
 
-export const dynamic = 'force-dynamic';
-
 export default async function NewReservationPage({ params }: { params: { slug: string } }) {
-  const me = await serverFetch('/api/auth/me');
-  if (me.status === 401) redirect(`/login?next=/groups/${params.slug}/reservations/new`);
+  const user = await getUserFromCookies();
+  if (!user) redirect(`/login?next=/groups/${params.slug}/reservations/new`);
   const res = await serverFetch(
     `/api/devices?groupSlug=${encodeURIComponent(params.slug)}`
   );
