+++ web/src/lib/auth/server.ts
@@ -1,0 +1,6 @@
+import 'server-only';
+import { readUserFromCookie } from '../auth';
+
+export async function getUserFromCookies() {
+  return readUserFromCookie();
+}
